+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="./public/logo/logo 1.png" />
+    <link rel="icon" type="image/svg+xml" href="/logo/logo 1.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>RunTheStack</title>
   </head>
