+++ src/Routes/Routes.jsx
@@ -1,19 +1,23 @@
-import { createBrowserRouter } from 'react-router-dom';
-import Main from '../Layouts/Main';
-import Home from '../Pages/Home/Home';
-
+import { createBrowserRouter } from "react-router-dom";
+import Main from "../Layouts/Main";
+import Home from "../Pages/Home/Home";
+import Courses from "../component/Courses/Courses";
 
 export const router = createBrowserRouter([
-    {
-        path: '/',
-        element: <Main></Main>,
-        children: [
-            {
-                path: '/',
-                element: <Home/>
-            },
-        ]
-    }
-])
+  {
+    path: "/",
+    element: <Main></Main>,
+    children: [
+      {
+        path: "/",
+        element: <Home />,
+      },
+      {
+        path: "/courses",
+        element: <Courses />,
+      },
+    ],
+  },
+]);
 
-export default router
+export default router;
