+++ src/component/CourseDetails/CourseDetails.jsx
@@ -1,0 +1,1 @@
+
